requirements.txt
langgraph>=0.0.64
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
elevenlabs>=0.2.0
